requirements.txt
# Dependencias de Token Helper
# Instalar con: pip install -r requirements.txt

# JDBC para Python
jaydebeapi>=1.2.3

# Variables de entorno
python-dotenv>=1.0.0

# Nota: También necesitas:
# - JDK/JRE instalado en el sistema
# - jtds-1.3.1.jar en la raíz del proyecto

